+++ server/hub.go
@@ -60,6 +60,7 @@ func (h *Hub) run() {
 	h.unregisterChan = make(chan *connection)
 	h.readChan = make(chan *connection)
 	h.writeChan = make(chan interface{}, 256)
+
 	for {
 		select {
 		case o := <-h.writeChan:
